+++ packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @comet/eslint-config
 
+## 8.1.1
+
+### Patch Changes
+
+- @comet/eslint-plugin@8.1.1
+
+## 8.1.0
+
+### Patch Changes
+
+- @comet/eslint-plugin@8.1.0
+
 ## 8.0.0
 
 ### Major Changes